backend/backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
python-multipart
httpx
python-dotenv
authlib
